Split views in `mane` into their own file.
